Package ounit2

Sub-indexes

advanced

Library ounit2

This library exposes the following toplevel modules:

OUnit Unit test building blocks

OUnit2 Unit test building blocks (v2).

Library ounit2.advanced

This library exposes the following toplevel modules:

OUnitAssert

OUnitBracket

OUnitCache

OUnitCheckEnv

OUnitChooser Heuristic to pick a test to run.

OUnitConf

OUnitCore

OUnitDiff Unit tests for collection of elements

OUnitLogger

OUnitLoggerCI

OUnitLoggerHTML

OUnitLoggerHTMLData

OUnitLoggerJUnit

OUnitLoggerStd

OUnitPlugin Standard functions for plugin (register, choose).

OUnitPropList Property list.

OUnitResultSummary

OUnitRunner

OUnitRunnerProcesses Use processes to run several tests in parallel. * * Run processes that handle running tests. The processes read test, execute * it, and communicate back to the master the log. * * This need to be done in another process because OCaml Threads are not truly * running in parallel. Moreover we cannot use Unix.fork because it's not * portable

OUnitShared

OUnitState Manipulate the state of OUnit runner.

OUnitTest

OUnitTestData

OUnitUtils Utilities for OUnit