Packages located relative to stdlib

Libraries

The following libraries are found in this directory.

stdlib

Modules

The following is a list of all of the modules found at this filesystem path.

Arith_status Flags that control rational arithmetic.

Big_int Operations on arbitrary-precision integers.

CamlinternalFormat

CamlinternalFormatBasics

CamlinternalLazy Run-time support for lazy values. All functions in this module are for system use only, not for the casual user.

CamlinternalMod Run-time support for recursive modules. All functions in this module are for system use only, not for the casual user.

CamlinternalOO Run-time support for objects and classes. All functions in this module are for system use only, not for the casual user.

Nat

Num Operation on arbitrary-precision numbers.

Ratio Operation on rational numbers.

Std_exit

Stdlib The OCaml Standard library.